+++ projects/js-packages/svelte-data-sync-client/src/SyncedStore.ts
@@ -1,20 +1,28 @@
-import { writable } from 'svelte/store';
+import deepEqual from 'deep-equal';
+import { Writable, writable } from 'svelte/store';
+import { ApiError } from './ApiError';
+import {
+	Pending,
+	SyncedStoreInterface,
+	SyncedWritable,
+	SyncedStoreCallback,
+	SyncedStoreError,
+} from './types';
 import { sleep } from './utils';
-import type { Pending, SyncedStoreInterface, SyncedWritable, SyncedStoreCallback } from './types';
 
 /*
  * A custom Svelte Store that's used to indicate if a value is being synced.
  */
 export class SyncedStore< T > {
 	private store: SyncedWritable< T >;
+	private errorStore: Writable< SyncedStoreError< T >[] >;
 	private pending: Pending;
-	private pendingValue: T | undefined;
-	private requestLock = false;
-	private failedToSync = Symbol( 'failedToSync' );
 	private updateCallback?: SyncedStoreCallback< T >;
+	private abortController: AbortController;
 
 	constructor( initialValue?: T ) {
 		this.store = this.createStore( initialValue );
+		this.errorStore = writable< SyncedStoreError< T >[] >( [] );
 		this.pending = this.createPendingStore();
 	}
 
@@ -22,13 +30,26 @@ export class SyncedStore< T > {
 		const store = writable< T >( initialValue );
 
 		// Send the store value to the API
-		store.set = value => {
+		const set = value => {
 			store.update( prevValue => {
 				// Synchronize is an async function, but is called without await here.
 				// This intentionally prevents the store from waiting for the request to complete.
 				// This is because we want the store to update immediately,
 				// and then the request to happen in the background.
-				this.debouncedSynchronize( prevValue, value );
+				this.abortableSynchronize( prevValue, value );
+				return value;
+			} );
+		};
+
+		type SvelteUpdater = typeof store.update;
+		const update: SvelteUpdater = updateCallback => {
+			store.update( prevValue => {
+				// Structured Clone is necessary because
+				// the updateCallback function may mutate the value
+				// And debouncedSynchronize may fail an object comparison
+				// because of it.
+				const value = updateCallback( prevValue );
+				this.abortableSynchronize( prevValue, value );
 				return value;
 			} );
 		};
@@ -38,118 +59,95 @@ export class SyncedStore< T > {
 		};
 
 		return {
-			...store,
+			subscribe: store.subscribe,
+			set,
+			update,
 			override,
 		};
 	}
 
-	public getPublicInterface(): SyncedStoreInterface< T > {
-		return {
-			store: this.store,
-			pending: {
-				subscribe: this.pending.subscribe,
-			},
-			setCallback: this.setCallback.bind( this ),
-		};
-	}
-
-	public setCallback( callback: SyncedStoreCallback< T > ) {
+	/**
+	 * A callback that will synchronize the store in some way.
+	 * By default, this is set to endpoint.POST in the client initializer
+	 */
+	private setCallback( callback: SyncedStoreCallback< T > ) {
 		this.updateCallback = callback;
 	}
 
-	private async synchronize( prevValue, value: T ): Promise< T | typeof this.failedToSync > {
+	/**
+	 * Attempt to synchronize the store with the API.
+	 */
+	private async synchronize( value: T ): Promise< T | ApiError > {
 		if ( ! this.updateCallback ) {
 			return value;
 		}
-		const result = await this.updateCallback( value );
-
-		// Success is only when the updateCallback result matches the value.
-		if ( this.equals( result, value ) ) {
-			return result;
-		}
-
-		return this.failedToSync;
-	}
-
-	private async debouncedSynchronize( prevValue, value: T, attempt = 0 ) {
-		// attempt = 0 indicates that this is run from the store.set() function.
-		const syncRequestFromStore = attempt === 0;
-
-		// If this is an update issued by the store, then there are two scenarios,
-		// where we don't need to issue a new Synchronize request:
-		// 1. The previous value is the same as the new value.
-		// 2. There's already a value pending, and it's the same as the new value.
-		//
-		if (
-			syncRequestFromStore &&
-			( this.equals( prevValue, value ) ||
-				( this.pendingValue !== undefined && this.equals( this.pendingValue, value ) ) )
-		) {
-			return;
-		}
-
-		// If this is an update from the store, then we need to update the pending value.
-		// Store the value in case the UI changes rapidly.
-		// This will be read after the debounce.
-		if ( syncRequestFromStore ) {
-			this.pendingValue = value;
-		}
-
-		// Prevent multiple requests from being sent at once.
-		if ( this.requestLock && syncRequestFromStore ) {
-			return;
-		}
-		this.requestLock = true;
-
-		// If UI Changes rapidly, wait for it to settle before issuing the request.
-		await sleep( 200 + 600 * attempt );
-
-		// The pending value may change while debouncing.
-		// If it's now the same as the previous value, don't send the request.
-		if ( this.pendingValue === prevValue ) {
-			// eslint-disable-next-line no-console
-			this.requestLock = false;
-			this.pendingValue = undefined;
-			return;
-		}
 
 		try {
-			const result = await this.synchronize( prevValue, this.pendingValue );
-			// Retry on two conditions:
-			// 1) The request failed to sync, but didn't throw an error
-			// 2) The pending value has changed since the request was sent
-			if ( result === this.failedToSync && attempt < 3 ) {
-				return this.debouncedSynchronize( prevValue, this.pendingValue, attempt + 1 );
+			const result = await this.updateCallback( value, this.abortController.signal );
+
+			// Success is only when the updateCallback result matches the value.
+			if ( this.equals( result, value ) ) {
+				return result ? result : value;
+			}
+		} catch ( error ) {
+			if ( error instanceof ApiError || error.name === 'ApiError' ) {
+				return error as ApiError;
 			}
 
-			if ( syncRequestFromStore && result !== this.failedToSync && result !== this.pendingValue ) {
-				// Using result instead of prevValue here
-				// Because we received fresher "real value" from the API.
-				return this.debouncedSynchronize( result, this.pendingValue, 1 );
-			}
-
-			if ( result !== this.failedToSync ) {
-				this.requestLock = false;
-				this.pendingValue = undefined;
-				return result;
-			}
-		} catch ( e ) {
-			// eslint-disable-next-line no-console
-			console.error( e );
+			// Rethrow the error if it's not an ApiError.
+			throw error;
 		}
 
-		// At this point, the request has either thrown and logged an error
-		// or the value returned from the API doesn't match the value sent to the API.
-		// Flip back the value and allow the synchronization to happen again.
-		this.store.override( prevValue );
-		this.requestLock = false;
-		this.pendingValue = undefined;
+		return new ApiError( 'SyncedStore::synchronize', 'failed_to_sync', 'Failed to sync' );
+	}
 
-		// eslint-disable-next-line no-console
-		console.error(
-			"Failed to sync value to the API. The value returned from the API doesn't match the value sent to the API."
-		);
-		return this.failedToSync;
+	/**
+	 * A debounced version of synchronize.
+	 * This is used to prevent the API from being spammed with requests.
+	 * It also prevents the store from updating when the API returns an error.
+	 */
+	private async abortableSynchronize( prevValue: T, value: T, retry = 0 ) {
+		if ( this.abortController ) {
+			this.abortController.abort();
+		}
+
+		this.abortController = new AbortController();
+		const signal = this.abortController.signal;
+
+		await sleep( 200 + 600 * retry );
+		if ( signal.aborted ) {
+			return;
+		}
+		const result = await this.synchronize( value );
+		if ( signal.aborted ) {
+			return;
+		}
+
+		if ( result instanceof ApiError ) {
+			if ( retry < 3 ) {
+				// Wait a second before retrying.
+				await sleep( 1000 );
+				if ( signal.aborted ) {
+					return;
+				}
+				this.abortableSynchronize( prevValue, value, retry + 1 );
+				return;
+			}
+			this.errorStore.update( errors => {
+				errors.push( {
+					time: Date.now(),
+					previousValue: prevValue,
+					value,
+					location: result.location,
+					status: result.status,
+					message: result.message,
+				} );
+				return errors;
+			} );
+			this.store.override( prevValue );
+		}
+
+		return result;
 	}
 
 	private createPendingStore(): Pending {
@@ -163,8 +161,27 @@ export class SyncedStore< T > {
 
 	private equals( a: unknown, b: unknown ) {
 		if ( typeof a === 'object' && typeof b === 'object' ) {
-			return Object.entries( a ).sort().toString() === Object.entries( b ).sort().toString();
+			return deepEqual( a, b );
 		}
+
 		return a === b;
+	}
+
+	/**
+	 * All of the class methods in this class are private.
+	 * Use this method to get the public interface of this class,
+	 * exposing as little as possible.
+	 */
+	public getPublicInterface(): SyncedStoreInterface< T > {
+		return {
+			store: this.store,
+			pending: {
+				subscribe: this.pending.subscribe,
+			},
+			errors: {
+				subscribe: this.errorStore.subscribe,
+			},
+			setCallback: this.setCallback.bind( this ),
+		};
 	}
 }