+++ projects/plugins/jetpack/extensions/blocks/subscriptions/panel.js
@@ -16,9 +16,23 @@ import './panel.scss';
 import { META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS } from './constants';
 import { NewsletterAccess, accessOptions, MisconfigurationWarning } from './settings';
 import { isNewsletterFeatureEnabled } from './utils';
+
+function AccessLevelSelectorPanel( { setPostMeta, accessLevel } ) {
+	if ( ! isNewsletterFeatureEnabled() ) {
+		return null;
+	}
+
+	return (
+		<PluginDocumentSettingPanel title={ __( 'Newsletter', 'jetpack' ) }>
+			<NewsletterAccess setPostMeta={ setPostMeta } accessLevel={ accessLevel } />
+		</PluginDocumentSettingPanel>
+	);
+}
+
 export default function SubscribePanels() {
 	const [ subscriberCount, setSubscriberCount ] = useState( null );
-	const [ postMeta = [], setPostMeta ] = useEntityProp( 'postType', 'post', 'meta' );
+	const postType = useSelect( select => select( editorStore ).getCurrentPostType(), [] );
+	const [ postMeta = [], setPostMeta ] = useEntityProp( 'postType', postType, 'meta' );
 
 	const accessLevel =
 		postMeta[ META_NAME_FOR_POST_LEVEL_ACCESS_SETTINGS ] ?? Object.keys( accessOptions )[ 0 ];
@@ -36,26 +50,29 @@ export default function SubscribePanels() {
 	const showMisconfigurationMessage = postVisibility !== 'public' && accessLevel !== 'everybody';
 
 	// Only show this for posts for now (subscriptions are only available on posts).
-	const postType = useSelect( select => select( editorStore ).getCurrentPostType(), [] );
 	const postWasEverPublished = useSelect(
 		select =>
 			select( editorStore ).getEditedPostAttribute( 'meta' )?.jetpack_post_was_ever_published,
 		[]
 	);
 
-	if ( 'post' !== postType ) {
-		return null;
-	}
-
-	// Subscriptions will not be triggered for a post that was already published in the past
-	if ( postWasEverPublished ) {
-		return null;
-	}
-
 	// Subscriptions will not be triggered on private sites (on WordPress.com simple and WoA),
 	// nor on sites that have not been launched yet.
 	if ( isPrivateSite() || isComingSoon() ) {
 		return null;
+	}
+
+	// Subscriptions are only available for posts. Additionally, we will allow access level selector for pages.
+	// TODO: Make it available for pages later.
+	if ( postType !== 'post' ) {
+		return null;
+	}
+
+	// Subscriptions will not be triggered for a post that was already published in the past and the email was sent.
+	// We still need to render the access level selector, as historical posts need a way to edit their access level for people visiting them on the web.
+	// TODO: Additionally, pages also can be protected. They will not send an email, but can be a resource that needs the acces selector.
+	if ( postWasEverPublished ) {
+		return <AccessLevelSelectorPanel setPostMeta={ setPostMeta } accessLevel={ accessLevel } />;
 	}
 
 	// Do not show any panels when we have no info about the subscriber count, or it is too low.
@@ -70,12 +87,7 @@ export default function SubscribePanels() {
 
 	return (
 		<>
-			{ isNewsletterFeatureEnabled() && (
-				<PluginDocumentSettingPanel title={ __( 'Newsletter', 'jetpack' ) }>
-					<NewsletterAccess setPostMeta={ setPostMeta } accessLevel={ accessLevel } />
-				</PluginDocumentSettingPanel>
-			) }
-
+			<AccessLevelSelectorPanel setPostMeta={ setPostMeta } accessLevel={ accessLevel } />
 			<PluginPrePublishPanel
 				className="jetpack-subscribe-pre-publish-panel"
 				initialOpen