+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.1] - 2023-03-08
+### Changed
+- Updated package dependencies. [#29216]
+
+## [0.17.0] - 2023-02-28
+### Added
+- Add Social Image Generator editor panel to post sidebar [#28737]
+- Add Social Image Generator feature flag to Jetpack Social [#29001]
+- Jetpack Social: Add Mastodon and default media upload restrictions [#29034]
+
+### Removed
+- Removed default image for SIG as it's not used yet [#29206]
+
+### Fixed
+- Update React peer dependencies to match updated dev dependencies. [#28924]
+
 ## [0.16.1] - 2023-02-20
 ### Changed
 - Minor internal updates.
@@ -201,6 +217,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[0.17.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.17.0...v0.17.1
+[0.17.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.15.2...v0.16.0
 [0.15.2]: https://github.com/Automattic/jetpack-publicize-components/compare/v0.15.1...v0.15.2